No functionality change; code cleanup
CMS/OABA/Transitivity/URM/E2 API/EJB and related classes

+++ 2.7.x/choicemaker-cm/choicemaker-j2ee/com.choicemaker.cm.batch.ejb/src/main/java/com/choicemaker/cm/batch/ejb/BatchJobUtils.java
@@ -13,7 +13,7 @@ public class BatchJobUtils {
 
 	public static void updateBatchJobStatusFromTo(final BatchJobEntity batchJob,
 			final BatchJobStatus current, final BatchJobStatus next) {
-		updateBatchJobStatusFromTo(batchJob, current, next, (int) 0);
+		updateBatchJobStatusFromTo(batchJob, current, next, 0);
 	}
 
 	public static final int MAX_RETRY = 5;
